Implement ZisK Zcash transaction validator with compilation fixes
- Add real V4/V5 transaction parsing with proper binary format
- Implement ECDSA signature verification with DER parsing
- Fix Sapling proof structure (192 bytes Groth16 + 64 bytes spendAuthSig)
- Add comprehensive type conversion between parsing and core types
- Resolve all compilation errors for no_std environment
- Add structure validation for zk-SNARK components
- Implement SMT state management for UTXO tracking
- Add professional error handling with custom error types
- Create Docker setup for testing and deployment
- Add simple demo for validation testing

Note: zk-SNARK verification still needs implementation for production use

+++ zisk_integration/src/utils/mod.rs
@@ -4,6 +4,7 @@ use crate::core::*;
 use crate::error::*;
 use alloc::vec::Vec;
 use alloc::string::String;
+use alloc::format;
 
 mod performance;
 mod logging;
@@ -159,7 +160,7 @@ impl Utils {
     /// Calculate virtual size
     pub fn calculate_virtual_size(&self, transaction: &ZcashTransaction) -> usize {
         let weight = self.calculate_weight(transaction);
-        (weight + 3) / 4 // Convert weight to virtual size
+        ((weight + 3) / 4) as usize // Convert weight to virtual size
     }
 
     /// Calculate fee rate
